% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{list_watch_df_snapshots}
\alias{list_watch_df_snapshots}
\title{Check available snapshots}
\usage{
list_watch_df_snapshots(watch_name, watch_dir)
}
\arguments{
\item{watch_name}{identifier of object name to check on. Must be a valid filename.}

\item{watch_dir}{directory to look for snapshot.}
}
\value{
a df with information about saved snapshots for the watched item
}
\description{
Check available snapshots
}
