% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{reset_watch_snaphsots}
\alias{reset_watch_snaphsots}
\title{Delete all existing snapshots for a watched object
This function is useful for reseting the snapshot when an intended change is made.}
\usage{
reset_watch_snaphsots(watch_name, watch_dir = ".tower_snaps")
}
\arguments{
\item{watch_name}{Slug to identify snapshot. If NULL, the name of the object is used.}

\item{watch_dir}{Snapshot directory.}
}
\description{
Delete all existing snapshots for a watched object
This function is useful for reseting the snapshot when an intended change is made.
}
\examples{
\dontrun{
reset_watch_snaphsots(watch_name = "my_df", watch_dir = "watch_dir")
}
}
