% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watch_df}
\alias{watch_df}
\title{Quick and dirty snapshot testing for a dataframe}
\usage{
watch_df(
  df,
  watch_name = NULL,
  watch_dir = "tower_snaps",
  if_diff = c("error", "warn", "silent")
)
}
\arguments{
\item{df}{Dataframe to save a snapshot of.}

\item{watch_name}{Identifier of object name to check on. Must be a valid filename.}

\item{watch_dir}{Directory to look for snapshot.}

\item{if_diff}{Desired behaviour if dataframes are not the same.}
}
\value{
boolean, true if no snapshot or snapshot is identical.
}
\description{
First, this function checks if a snapshot exists for the dataframe passed.
If no snapshot exists, it creates a snapshot in watch_dir.
If a snapshot does exists, it compares the snapshot to the passed dataframe using all.equal
with the caveat that it ignores differences in row or column order.
}
