% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{list_watch_df_snapshots}
\alias{list_watch_df_snapshots}
\title{Check available snapshots}
\usage{
list_watch_df_snapshots(watch_name, watch_dir = ".tower_snaps")
}
\arguments{
\item{watch_name}{Slug to identify snapshot. If NULL, the name of the object is used.}

\item{watch_dir}{Snapshot directory.}
}
\value{
a df with information about saved snapshots for the watched item.
}
\description{
Check available snapshots
}
\examples{
\dontrun{
list_watch_df_snapshots(watch_name = "my_df", watch_dir = ".tower_snaps")
}
}
