% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{reset_watch_snaphsots}
\alias{reset_watch_snaphsots}
\title{Delete all existing snapshots for a watched object
This function is useful for reseting the snapshot when an intended change is made.}
\usage{
reset_watch_snaphsots(watch_name, watch_dir)
}
\arguments{
\item{watch_name}{Identifier of object name to check on. Must be a valid filename.}

\item{watch_dir}{Directory to look for snapshot.}
}
\description{
Delete all existing snapshots for a watched object
This function is useful for reseting the snapshot when an intended change is made.
}
