% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watch_df}
\alias{watch_df}
\title{Create a snapshot of a dataframe or check against an existing snapshot}
\usage{
watch_df(
  df,
  watch_name = NULL,
  watch_dir = ".tower_snaps",
  if_diff = c("error", "warn", "silent")
)
}
\arguments{
\item{df}{Dataframe to snapshot test.}

\item{watch_name}{Slug to identify snapshot. If NULL, the name of the object is used.}

\item{watch_dir}{Snapshot directory.}

\item{if_diff}{Desired behaviour if dataframes are not the same.}
}
\value{
boolean, true if no snapshot or snapshot is identical.
}
\description{
If a snapshot exists with the passed watch_name in watch_dir, this function compares the snapshot to the passed dataframe using all.equal
after standardizing row and column order.
}
\examples{
my_df <- data.frame("numeric_col" = c(1, 2))
# Fancy code you want to refactor that does stuff to my_df goes here
# Saves first time it's run, returns true afterwards
watch_dir = tempdir()
watch_df(my_df, watch_dir = watch_dir)
my_df <- dplyr::mutate(my_df, numeric_col = numeric_col + 1) 
# Returns false
 watch_df(
   my_df,
   watch_name = "my_df",
   watch_dir = watch_dir,
   if_diff = "silent"
 )
}
