% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watch_df}
\alias{watch_df}
\title{Quick and dirty snapshot testing for a dataframe}
\usage{
watch_df(
  df,
  watch_name = NULL,
  watch_dir = "tower_snaps",
  if_diff = c("error", "warn", "silent")
)
}
\arguments{
\item{df}{Dataframe to save a snapshot of.}

\item{watch_name}{Identifier of object name to check on. Must be a valid filename.}

\item{watch_dir}{Directory to look for snapshot.}

\item{if_diff}{Desired behaviour if dataframes are not the same.}
}
\value{
boolean, true if no snapshot or snapshot is identical.
}
\description{
Checks the watch_dir for the passed watch name and comapres it against the df or saves a new 
snapshot if one is not available.
}
